ai-service/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Basic data processing
numpy==1.24.3
pandas==2.0.3

# HTTP client
httpx==0.25.2
requests==2.31.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
